Quizzes Creation and Taking
I made the quizzes dynamic and be able to have the 3 types like on efundi, long form, multiple choice, fill in the blank. However in doing this I do realize we need to update the database a little bit to accommodate the mutli-choice questions. Also added quiz description which is basically where a lecturer would say the what it's about and the no cheating stuff.

I think it should work, didn't really have a way to test it, just keep me updated and let me know if I need to change things.