Quizzes Overhaul
Updated the Database to be able to accommodate the 3 question types.

Redid all the Quiz codes to not try and use columns outside the database framework. Improved on the UI to better be able to use and obtain relevant information for the database.

Completely changed how the taking a quiz works. The main Take quiz will now only ask you to select a module and quiz and from there I added 3 answer forms to be called as needed for each question type.

The responses table is mostly unchanged, except I set the answer to MAX instead of 225 to be able to receive a long form answer.

The questions happen sequentially MCQ then FBQ then LFQ. Always in that order, depending on what questions exist for a particular quiz.

At the end of a quiz whether completed, ran out of time, or due date passed. The score for MCQ and FBQ is auto calculated. LFQ marks are added to total, but a label will notify that long form is marked manually. Gradebook is updated accordingly and the comment is filled with the message if long form needs to be marked.